+++ projects/amdsmi/amdsmi_cli/amdsmi_parser.py
@@ -1820,6 +1820,7 @@ class AMDSMIParser(argparse.ArgumentParser):
         throttle_help = (
             "Displays throttle accumulators;\n    Only available for MI300 or newer ASICs"
         )
+        partition_help = "Displays XCP partition metrics"
 
         # Help text for Arguments only on Hypervisors
         schedule_help = "All scheduling information"
@@ -1988,6 +1989,9 @@ class AMDSMIParser(argparse.ArgumentParser):
                     action="store_true",
                     required=False,
                     help=argparse.SUPPRESS,
+                )
+                metric_parser.add_argument(
+                    "-X", "--partition", action="store_true", required=False, help=partition_help
                 )
 
             # Options to only display to Hypervisors